requirements.txt
# requirements.txt
streamlit>=1.28.0
mistralai>=0.1.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
nltk>=3.8.0
matplotlib>=3.7.0
plotly>=5.17.0
joblib>=1.3.0